IMPORTANT - READ DESCRIPTION
Added a git dependancy package for the base of the node editor for the conversation tree. If you  use UNITY HUB, update to the latest version. There is a bug in the older version that means you cannot find the git PATH.
If you cannot find the git PATH with the latest version (at the moment 2.2.2) then you will have to install Git on your machine and maybe tell the system the path (turtorial here: https://stackoverflow.com/questions/26620312/installing-git-in-path-with-github-client-for-windows/34767523#34767523
)