Second Iteration of Fish fights, other fishing logic and vibrations
Finished the second iteration of fish fighting. Noticed a couple bugs that will get ironed out tomorrow.

Made a few changes to the logic of the fish so the fish are faster when attracted

Added vibration to the controller for when the fish is attracted or interacting. This makes the player know when they should stop  reeling and wait