Began constructing a solid method for conversation manager. Also made a tween manager
decided on creating an event system for the story/convo manager, allowing things like the UI to subscribe to certain events, which organises the code much more nicely
also made a generalised tween manager. It allows the creation of a path in code which will then be followed by the reference variable given to it. The full tween can be made up of many composite tweens, giving a lot of options for changing things like positions, scales, alpha, etc. in the same full tween. Once a full tween is finished, the passed in delegate is invoked to signal the end of the tween